step3 test: 리팩토링한 부분 테스트 수정
