+++ src/main/kotlin/racingWinner/README.md
@@ -11,7 +11,7 @@
    3. 전진한 자동차의 position을 1 증가
 
 ## 기능 목록 
-- [ ] 자동차는 이름을 가진다.
-- [ ] 자동차의 이름은 5자를 초과할 수 없음
-- [ ] 자동차는 4 이상일 경우 움직인다
-- [ ] 자동차는 4 미만인 경우 멈춰있는다
+- [x] 자동차는 이름을 가진다.
+- [x] 자동차의 이름은 5자를 초과할 수 없음
+- [x] 자동차는 4 이상일 경우 움직인다
+- [x] 자동차는 4 미만인 경우 멈춰있는다
